# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/categories/region/{region_code}": {
            "get": {
                "description": "根据地区编码获取该地区的分类列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "获取地区分类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地区编码",
                        "name": "region_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/regions": {
            "get": {
                "description": "获取所有地区列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取地区列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的地区（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "创建地区",
                "parameters": [
                    {
                        "description": "地区信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Region"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或无效的地区级别",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "409": {
                        "description": "地区代码已存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/code/{code}": {
            "get": {
                "description": "根据地区编码获取地区信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "根据编码获取地区",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地区编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "地区不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/heritage-centers": {
            "get": {
                "description": "获取所有非遗保护中心地区",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取非遗保护中心",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/level/{level}": {
            "get": {
                "description": "根据级别获取地区列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取指定级别地区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地区级别",
                        "name": "level",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的地区级别",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/parent/{parent_id}": {
            "get": {
                "description": "根据父地区ID获取子地区列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取子地区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "父地区ID",
                        "name": "parent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的地区ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/root": {
            "get": {
                "description": "获取一级地区（无父级的地区）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取根地区",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/{id}": {
            "get": {
                "description": "根据ID获取地区信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取地区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的地区ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "地区不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新地区信息（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "更新地区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "地区信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Region"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "地区不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "409": {
                        "description": "地区代码已存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除地区（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "删除地区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或无法删除有子节点的地区",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "地区不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/regions/{id}/with-children": {
            "get": {
                "description": "根据ID获取地区信息及其所有子地区",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "获取地区及子地区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的地区ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "地区不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backend_internal_models.Region": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "行政区划代码（如 110101）",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "cultureDesc": {
                    "description": "地域文化背景描述",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "isHeritageCenter": {
                    "description": "是否为非遗重点保护区域",
                    "type": "boolean"
                },
                "level": {
                    "description": "级别：省/市/县",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentID": {
                    "description": "上级区域 ID",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "backend_pkg_response.ErrorField": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "backend_pkg_response.PageInfo": {
            "type": "object",
            "properties": {
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/backend_pkg_response.ErrorField"
                },
                "page": {
                    "$ref": "#/definitions/backend_pkg_response.PageInfo"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}